:recycle: Refactor api hooks
